# openapi.yaml
openapi: 3.1.0
info:
  title: ClassnoteX API
  description: Backend API for ClassnoteX - AI-powered recording and transcription
    app.
  version: 0.4.0
paths:
...
  /sessions:
    post:
      tags:
      - Sessions
      summary: Create Session
      operationId: create_session_sessions_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Sessions
      summary: List Sessions
      operationId: list_sessions_sessions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
                title: Response List Sessions Sessions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /admin/overview:
    get:
      tags:
      - Admin
      - Admin
      summary: Get Admin Overview
      description: 'Get KPI overview for admin dashboard.

        TODO: Add proper admin role check.'
      operationId: get_admin_overview_admin_overview_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewResponse'
      security:
      - OAuth2PasswordBearer: []
  /admin/users:
    get:
      tags:
      - Admin
      - Admin
      summary: Get Admin Users
      description: Get paginated user list with filters.
      operationId: get_admin_users_admin_users_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: query
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by name, email, or UID
          title: Query
        description: Search by name, email, or UID
      - name: plan
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by plan
          title: Plan
        description: Filter by plan
      - name: active
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filter by activity: 7d, 30d, inactive'
          title: Active
        description: 'Filter by activity: 7d, 30d, inactive'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
                title: Response Get Admin Users Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/users/rankings:
    get:
      tags:
      - Admin
      - Admin
      summary: Get User Rankings
      description: Get top users by specified metric.
      operationId: get_user_rankings_admin_users_rankings_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: metric
        in: query
        required: false
        schema:
          type: string
          description: 'Metric: recording, cost, sessions'
          default: recording
          title: Metric
        description: 'Metric: recording, cost, sessions'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingItem'
                title: Response Get User Rankings Admin Users Rankings Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/users/{uid}:
    get:
      tags:
      - Admin
      - Admin
      summary: Get User Detail
      description: Get detailed user info with usage stats.
      operationId: get_user_detail_admin_users__uid__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
          title: Uid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/config/pricing:
    get:
      tags:
      - Admin
      - Admin
      summary: Get Pricing Config
      description: Get current pricing configuration.
      operationId: get_pricing_config_admin_config_pricing_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingConfig'
      security:
      - OAuth2PasswordBearer: []
    patch:
      tags:
      - Admin
      - Admin
      summary: Update Pricing Config
      description: Update pricing configuration.
      operationId: update_pricing_config_admin_config_pricing_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PricingConfig'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /admin/sessions:
    get:
      tags:
      - Admin
      - Admin
      summary: Get Admin Sessions
      description: Get paginated session list with filters.
      operationId: get_admin_sessions_admin_sessions_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by status
          title: Status
        description: Filter by status
      - name: mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by mode
          title: Mode
        description: Filter by mode
      - name: userId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by userId
          title: Userid
        description: Filter by userId
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminSessionListItem'
                title: Response Get Admin Sessions Admin Sessions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateSessionRequest:
      properties:
        title:
          type: string
          title: Title
        mode:
          type: string
          title: Mode
          default: lecture
        userId:
          anyOf:
          - type: string
          - type: 'null'
          title: Userid
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        startAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startat
        endAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Endat
        syncToGoogleCalendar:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Synctogooglecalendar
          default: false
        visibility:
          type: string
          title: Visibility
          default: private
        transcriptionMode:
          anyOf:
          - $ref: '#/components/schemas/TranscriptionMode'
          - type: 'null'
          default: cloud_google
      type: object
      required:
      - title
      title: CreateSessionRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OverviewResponse:
      properties:
        total_users:
          type: integer
          title: Total Users
        dau:
          type: integer
          title: Dau
        wau:
          type: integer
          title: Wau
        mau:
          type: integer
          title: Mau
        new_users_today:
          type: integer
          title: New Users Today
        new_users_7d:
          type: integer
          title: New Users 7D
        total_sessions_today:
          type: integer
          title: Total Sessions Today
        total_sessions_7d:
          type: integer
          title: Total Sessions 7D
        total_recording_sec_today:
          type: number
          title: Total Recording Sec Today
        total_recording_sec_7d:
          type: number
          title: Total Recording Sec 7D
        estimated_cost_today_usd:
          type: number
          title: Estimated Cost Today Usd
        estimated_cost_7d_usd:
          type: number
          title: Estimated Cost 7D Usd
        jobs_failed_24h:
          type: integer
          title: Jobs Failed 24H
      type: object
      required:
      - total_users
      - dau
      - wau
      - mau
      - new_users_today
      - new_users_7d
      - total_sessions_today
      - total_sessions_7d
      - total_recording_sec_today
      - total_recording_sec_7d
      - estimated_cost_today_usd
      - estimated_cost_7d_usd
      - jobs_failed_24h
      title: OverviewResponse
...
    UserSummary:
      properties:
        uid:
          type: string
          title: Uid
        displayName:
          anyOf:
          - type: string
          - type: 'null'
          title: Displayname
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        plan:
          type: string
          title: Plan
          default: free
        providers:
          items:
            type: string
          type: array
          title: Providers
          default: []
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
        lastSeenAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lastseenat
        session_count_30d:
          type: integer
          title: Session Count 30D
          default: 0
        total_recording_min_30d:
          type: number
          title: Total Recording Min 30D
          default: 0.0
        total_recording_min_lifetime:
          type: number
          title: Total Recording Min Lifetime
          default: 0.0
        estimated_cost_30d_usd:
          type: number
          title: Estimated Cost 30D Usd
          default: 0.0
        active_badge:
          type: string
          title: Active Badge
          default: inactive
      type: object
      required:
      - uid
      title: UserSummary
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token